Use CC-BY as license
The Technical Charter recommends CC-BY, not Apache-2.0, as
the license for text. So let's use that for the documentation.
Code is still under Apache-2.0.

Since there is more than once license,
I put these licenses in a new LICENSES/ directory and named the files
using their SPDX identifiers along with their .txt extensions.
This is the convention recommended by the
[REUSE Specification](https://reuse.software/spec/).
Note that the CII Best Practices badge specifically notes this
as a common convention.

I modified the template to include a section with more license details.
The template now also includes a machine-readable identifier, as
explained at <https://spdx.dev/ids/>.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>