Add SECURITY.md file
There are scripts here, and it's conceivable they could have
a vulnerability. We *do* want *those* vulnerabilities reported to us.
However, we don't want people to report new/unknown vulnerabilities
in *other* projects here.

This adds a SECURITY.md file to hopefully make that clear, including
how to report vulnerabilities in our scripts.

I also tweaked the README, both to point to the SECURITY.md file, and to
make it clear what vulnerabilities we DO and DO NOT want reported.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>